Updated with most recent metadata/build date
Also updated the readme which contained an invalid path.

The version number of the gem in the source was not modified because
this was out of sync with the pushed meterpreter gem file anyway. The
version will remain 0.0.7 in source and the pushed meterpreter bins gem
will match this version.